backend/requirements.txt
fastapi[all]
uvicorn[standard]
motor
# passlib[bcrypt]
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
pydantic-settings
email-validator
pymongo
python-dotenv
beautifulsoup4
lxml